+++ source/compose-spec.ts
@@ -1,5 +1,4 @@
 import {add, subtract, compose} from 'rambda'
-// import { kebabCase } from 'string-fn';
 
 describe('compose', () => {
   it('happy', () => {
@@ -18,22 +17,4 @@ describe('compose', () => {
     )();
     result // $ExpectType void
   })
-
-  // TODO
-  /*
-  it('example', () => {
-    const label = compose(
-      kebabCase,
-      last,
-      split('/')
-    )(url)
-
-    const result = compose(
-      kebabCase,
-      (x: string[]) => x[0],
-      (x: string) => x.split('/')
-    )('url')
-    result // $ExpectType void
-  })
-  */
 })