+++ source/last-spec.ts
@@ -1,0 +1,19 @@
+import {last} from 'rambda'
+
+describe('R.last', () => {
+  it('string', () => {
+    const result = last('foo')
+    result // $ExpectType string
+  })
+
+  it('array', () => {
+    const result = last([1, 2, 3])
+    result // $ExpectType number | undefined
+  })
+
+  it('empty array', () => {
+    const result = last([])
+    result // $ExpectType undefined
+  })
+})
+